Test

See merge request jsql/jsql-api!80

+++ src/main/java/pl/jsql/api/dto/response/PaymentResponse.java
@@ -1,11 +1,9 @@
 package pl.jsql.api.dto.response;
 
-import java.math.BigDecimal;
-
 public class PaymentResponse {
 
     public Boolean active;
     public String plan;
-    public BigDecimal price;
+    public Double price;
 
 }
